refactor: Update QR service to use aggregate for retrieving QR data
Refactor the QR service to use the aggregate method for retrieving QR data. This change improves the efficiency and performance of the code by reducing the number of database queries. By using the aggregate method, we can fetch the necessary data in a single query, resulting in faster response times. This update enhances the overall functionality of the QR service and improves the user experience.