```text refactor: Add API_USER_EXIST function for checking if a user exists
Add a new API function, API_USER_EXIST, to check if a user exists. This function sends a GET request to the server endpoint '/users/users-exist' and returns the response data. By adding this function, we enhance the functionality of the application by providing a way to verify the existence of a user.