chore: Update npm dependencies for lazy loading and user management data table
This commit updates the npm dependencies in the client package.json and package-lock.json files. It adds the "ldrs" and "lottie-react" packages to enable lazy loading of images and improve the visual appeal of the SignUp and Home pages. Additionally, it includes the necessary dependencies for the user management data table in the Management page. These updates enhance the functionality and user experience of the application.