feat: Add toast notification for login error
This commit adds a toast notification to display an error message when there is an issue with the login process. It uses the "sonner" library to show the toast with a description indicating that the user needs to login again. This enhancement improves the user experience by providing clear feedback and guidance in case of login errors.