refactor: Update error message for existing QR in InsertQr function
The commit updates the error message in the InsertQr function of qr.service.ts. The message has been modified to provide more specific information by replacing "Qr already exist!" with "Id number already exist!". This refactor improves the clarity of the error message for users when attempting to insert a duplicate QR.