refactor: Update Mongoose connection URI to use environment variable
The commit updates the Mongoose connection URI in the app.module.ts file to use the environment variable `process.env.MONGODB_URI` instead of the hardcoded URI `mongodb://localhost:27017/upoints-kiosk`. This change allows for more flexibility in configuring the MongoDB connection and ensures that the application can connect to the appropriate database based on the environment.