refactor: Update favicon in index.html
The commit updates the favicon in the client/index.html file. It changes the href attribute of the link tag to point to a new logo image file, logo.jpg, located in the src/assets directory. This change ensures that the correct favicon is displayed for the website, improving the visual branding and user experience.